zap-app/requirements.txt
flask==2.3.2
opentelemetry-distro[otlp]
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk
requests